Update with interface changes
